split out one more function on grpc
