+++ python/ray/tests/test_actor_advanced.py
@@ -12,12 +12,12 @@ import ray.cluster_utils
 from ray._private.test_utils import (
     SignalActor,
     convert_actor_state,
-    kill_actor_and_wait_for_failure,
     make_global_state_accessor,
     run_string_as_driver,
     wait_for_condition,
     wait_for_pid_to_exit,
 )
+from ray._private.test_grpc_utils import kill_actor_and_wait_for_failure
 from ray._private.ray_constants import gcs_actor_scheduling_enabled
 from ray.experimental.internal_kv import _internal_kv_get, _internal_kv_put
