+++ python/ray/tests/test_reference_counting.py
@@ -14,10 +14,10 @@ import ray.cluster_utils
 from ray._private.test_utils import (
     SignalActor,
     convert_actor_state,
-    kill_actor_and_wait_for_failure,
     put_object,
     wait_for_condition,
 )
+from ray._private.test_grpc_utils import kill_actor_and_wait_for_failure
 
 logger = logging.getLogger(__name__)
